Extract the private fiber.h from fiber.cpp
Move the scheduler-internal declarations - the CQE tag constants, the
counter table, CxaEhGlobals, Fiber, CpuTimer, ProcessorState, and
SchedulerState - verbatim into src/fibers/fiber.h so further scheduler
work can split fiber.cpp by translation unit. No behavior change.
